+++ packages/stylelint-plugin/config/property.js
@@ -9,16 +9,16 @@ module.exports = {
   ],
   // Specify pattern of custom properties.
   'custom-property-pattern': null,
-  // Specify a blacklist of disallowed properties.
-  'property-blacklist': [],
+  // Specify a disallowed listed of disallowed properties.
+  'property-disallowed-list': [],
   // Specify lowercase or uppercase for properties.
   'property-case': 'lower',
   // Disallow unknown properties.
   'property-no-unknown': true,
   // Disallow vendor prefixes for properties.
   'property-no-vendor-prefix': true,
-  // Specify a whitelist of allowed properties.
-  'property-whitelist': null,
+  // Specify a allowed-list of allowed properties.
+  'property-allowed-list': null,
   // Disallow redundant values in shorthand properties.
   'shorthand-property-no-redundant-values': true,
 };